fix(leadsource): re-enable SMS channel encoding for edge submissions; extract shared logic
The channel-omission fix from f6c0267 was specifically to protect the legacy
Magento REST path from duplicate SMS subscriptions. It got applied blanket
across footer/modal (form.js) and (in this branch's earlier commit)
registration, even though registration and edge-locale footer/modal
submissions never touch that Magento endpoint. Re-enable channel encoding
wherever the submission actually goes to the new AEM Forms endpoint:

- blocks/form/form.js: footer/modal/minimized-modal now compute leadSource
  differently per branch -- getLeadSource() (channel-aware) for edge
  locales, getLegacyLeadSource() (always email-only, per the original fix)
  for locales still on Magento.
- widgets/forms/product-registration.js: registration has no legacy Magento
  path at all (decided last commit), so there's no duplicate-subscription
  risk -- restore SMS channel encoding (sub-emsms-reg-us). Still no
  SMS-only variant, since registration always treats email as opted in.
- widgets/account/account.js: unchanged behavior, just repointed to the new
  shared module.
- tests/pdp/integration.spec.js: updated expected leadSource values for the
  SMS-opted-in footer/modal cases to match.

Also moved the shared getLeadSource/getLegacyLeadSource helpers out of
scripts/scripts.js (loaded eagerly on every page) into a new
scripts/lead-source.js module, imported only by the handful of forms that
actually need it.
