+++ widgets/forms/product-registration.js
@@ -1,4 +1,5 @@
-import { getFormSubmissionUrl, getLocaleAndLanguage, getLeadSource } from '../../scripts/scripts.js';
+import { getFormSubmissionUrl, getLocaleAndLanguage } from '../../scripts/scripts.js';
+import { getLeadSource } from '../../scripts/lead-source.js';
 import getStatesProvincesOptions from './states-provinces.js';
 
 /**
@@ -166,9 +167,15 @@ export default async function decorate(widget) {
     const payload = Object.fromEntries(data.entries());
     payload.formId = `${locale}/${language}/product-registration`;
     payload.pageUrl = window.location.href;
-    // Registration's leadSource never varies by channel — SMS consent is carried solely
-    // by the separate smsOptIn field, matching production Magento's behavior.
-    payload.leadSource = getLeadSource('reg', (locale || 'us').toLowerCase());
+    // Registration submits only to the new AEM Forms endpoint (no legacy Magento path),
+    // so the channel-suppression concern that applies to Magento-backed forms doesn't
+    // apply here — SMS opt-in is safe to encode in the leadSource. Registration has no
+    // SMS-only path (email is always part of registering a product), so email is
+    // always treated as opted in; only the SMS checkbox varies the channel.
+    payload.leadSource = getLeadSource('reg', (locale || 'us').toLowerCase(), {
+      emailOptIn: true,
+      smsOptIn: payload.smsOptIn === 'yes',
+    });
 
     const submitButton = form.querySelector('button[type="submit"]');
     const buttonLabel = submitButton?.textContent;
